pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "check-physicell-it-startup"
version = "0.1.0"
description = "Monitor PhysiCell interactive tool startup on Galaxy"
requires-python = ">=3.11"
dependencies = [
    "bioblend>=1.2.0",
    "playwright>=1.40.0",
    "pytest>=8.0.0",
    "pytest-playwright>=0.5.0",
    "python-dotenv>=1.0.0",
    "pytest-timeout>=2.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--browser chromium --screenshot only-on-failure --tracing retain-on-failure"

[tool.setuptools.packages.find]
include = ["config*", "helpers*"]
